Add comment, use generic.
