+++ examples/practice-typescript/src/components/TableHeader /index.scss
@@ -5,5 +5,4 @@
   background: $color-bg-wrapper;
   min-height: 2rem;
   padding: 0.55rem 0 0.5rem 1.85rem;
-  top: 4.2rem;
 }
