+++ examples/practice-typescript/src/components/TableHeader /index.tsx
@@ -1,30 +1,12 @@
-import User from 'scripts/types/user';
+import { HeaderItem } from './HeaderItem';
 import './index.scss';
-import Product from 'scripts/types/product';
 
-interface HeaderItemProps {
-  label: string;
+interface HeaderItemsProps<T> {
+  headers: T[];
 }
 
-const HeaderItem = ({ label }: HeaderItemProps) => {
-  const formatLabel = (str: string) => {
-    return str
-      .replace(/\b\w/g, (char) => char.toUpperCase())
-      .replace(/([A-Z])/g, ' $1')
-      .trim();
-  };
-
-  return <th className="table-head">{formatLabel(label)}</th>;
-};
-
-interface TableHeaderProps {
-  headers: User[] | Product[];
-}
-
-const TableHeader = ({ headers = [] }: TableHeaderProps) => {
-  const headerKeys = Object.keys(headers[0]).filter(
-    (header) => header !== 'id'
-  );
+const TableHeader = <T,>({ headers }: HeaderItemsProps<T>) => {
+  const headerKeys: string[] = Object.keys(headers[0]);
 
   return (
     <tr>
